requirements.txt
argparse
pyserial
opencv-python
fer
tensorflow
openai
sounddevice
scipy
python-dotenv